# docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: Data Governance API
  version: 0.6.0
  description: API documentation for discovery, reporting, and platform operations.
servers:
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/v1/reporting/export/catalog.csv:
    get:
      summary: Export object catalog as CSV
      security:
        - bearerAuth: []
      responses:
        '200':
          description: CSV file download
  /api/v1/reporting/export/catalog.xlsx:
    get:
      summary: Export object catalog as Excel-compatible file
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Excel-compatible file download
  /api/v1/reporting/export/dependency/{objectId}.pdf:
    get:
      summary: Export dependency report PDF
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: objectId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PDF file download
        '404':
          description: Object not found
  /api/v1/reporting/share/visualization/{objectId}:
    post:
      summary: Create shareable visualization link
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: objectId
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                format:
                  type: string
                  enum: [svg, png]
                ttlMinutes:
                  type: integer
      responses:
        '201':
          description: Share link created
  /api/v1/reporting/schedules:
    get:
      summary: List report schedules
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Schedule list
    post:
      summary: Create report schedule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recipient]
              properties:
                recipient:
                  type: string
                type:
                  type: string
                  enum: [catalog, dependency]
                format:
                  type: string
                  enum: [csv, xlsx, pdf]
                frequency:
                  type: string
      responses:
        '201':
          description: Schedule created
  /api/v1/reporting/schedules/{scheduleId}/run:
    post:
      summary: Run schedule immediately
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: scheduleId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Schedule executed
        '404':
          description: Schedule missing
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT